Refactor order and Alpca config
Get rid of complted order and just let order have stock symbol attached
to it. Also remove alpaca config from the alpaca client. It is no longer
needed. Also remove some usused nuget packages.
