+++ AutomaticStockTrader.Domain/Domain.cs
@@ -11,14 +11,10 @@ namespace AutomaticStockTrader.Domain
 
     public record Order
     {
+        public string StockSymbol { get; init; }
         public DateTime OrderPlacedTime { get; init; }
         public long SharesBought { get; init; }
         public decimal MarketPrice { get; init; }
-    }
-
-    public record CompletedOrder : Order
-    {
-        public string StockSymbol { get; init; }
     }
 
     public record StrategysStock
