+++ AutomaticStockTrader.Repository/ITrackingRepository.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
@@ -7,7 +7,7 @@ namespace AutomaticStockTrader.Repository
     public interface ITrackingRepository : IDisposable
     {
         public Task AddPendingOrder(Domain.StrategysStock postion, Domain.Order order);
-        public Task CompleteOrder(Domain.CompletedOrder completedOrder);
+        public Task CompleteOrder(Domain.Order completedOrder);
         public Task<Domain.Position> GetOrCreateEmptyPosition(Domain.StrategysStock strategysStock);
         public IEnumerable<Domain.Order> GetCompletedOrders(Domain.StrategysStock strategysStock);
     }
